+++ ruma-identifiers/src/signing_key_id.rs
@@ -6,13 +6,13 @@ use ruma_identifiers_validation::{crypto_algorithms::SigningKeyAlgorithm, Error}
 
 /// Key identifiers used for homeserver signing keys.
 #[derive(Clone, Debug)]
-pub struct SigningKeyId {
+pub struct ServerSigningKeyId {
     full_id: Box<str>,
     colon_idx: NonZeroU8,
 }
 
-impl SigningKeyId {
-    /// Create a `SigningKeyId` from a `SigningKeyAlgorithm` and a `ServerId`.
+impl ServerSigningKeyId {
+    /// Create a `ServerSigningKeyId` from a `SigningKeyAlgorithm` and a `ServerId`.
     pub fn from_parts(algorithm: SigningKeyAlgorithm, version: &str) -> Self {
         let algorithm: &str = algorithm.as_ref();
 
@@ -25,7 +25,7 @@ impl SigningKeyId {
             NonZeroU8::new(algorithm.len().try_into().expect("no algorithm name len > 255"))
                 .expect("no empty algorithm name");
 
-        SigningKeyId { full_id: res.into(), colon_idx }
+        ServerSigningKeyId { full_id: res.into(), colon_idx }
     }
 
     /// Returns key algorithm of the server key ID.
@@ -39,15 +39,15 @@ impl SigningKeyId {
     }
 }
 
-fn try_from<S>(key_id: S) -> Result<SigningKeyId, Error>
+fn try_from<S>(key_id: S) -> Result<ServerSigningKeyId, Error>
 where
     S: AsRef<str> + Into<Box<str>>,
 {
     let colon_idx = ruma_identifiers_validation::signing_key_id::validate(key_id.as_ref())?;
-    Ok(SigningKeyId { full_id: key_id.into(), colon_idx })
+    Ok(ServerSigningKeyId { full_id: key_id.into(), colon_idx })
 }
 
-common_impls!(SigningKeyId, try_from, "Key ID with algorithm and version");
+common_impls!(ServerSigningKeyId, try_from, "Key ID with algorithm and version");
 
 #[cfg(test)]
 mod tests {
@@ -56,7 +56,7 @@ mod tests {
     #[cfg(feature = "serde")]
     use serde_json::{from_value as from_json_value, json, to_value as to_json_value};
 
-    use crate::{Error, SigningKeyId};
+    use crate::{Error, ServerSigningKeyId};
 
     #[cfg(feature = "serde")]
     use ruma_identifiers_validation::crypto_algorithms::SigningKeyAlgorithm;
@@ -64,7 +64,7 @@ mod tests {
     #[cfg(feature = "serde")]
     #[test]
     fn deserialize_id() {
-        let server_key_id: SigningKeyId = from_json_value(json!("ed25519:Abc_1")).unwrap();
+        let server_key_id: ServerSigningKeyId = from_json_value(json!("ed25519:Abc_1")).unwrap();
         assert_eq!(server_key_id.algorithm(), SigningKeyAlgorithm::Ed25519);
         assert_eq!(server_key_id.version(), "Abc_1");
     }
@@ -72,19 +72,23 @@ mod tests {
     #[cfg(feature = "serde")]
     #[test]
     fn serialize_id() {
-        let server_key_id: SigningKeyId = SigningKeyId::try_from("ed25519:abc123").unwrap();
+        let server_key_id: ServerSigningKeyId =
+            ServerSigningKeyId::try_from("ed25519:abc123").unwrap();
         assert_eq!(to_json_value(&server_key_id).unwrap(), json!("ed25519:abc123"));
     }
 
     #[test]
     fn invalid_version_characters() {
-        assert_eq!(SigningKeyId::try_from("ed25519:Abc-1").unwrap_err(), Error::InvalidCharacters);
+        assert_eq!(
+            ServerSigningKeyId::try_from("ed25519:Abc-1").unwrap_err(),
+            Error::InvalidCharacters
+        );
     }
 
     #[test]
     fn invalid_key_algorithm() {
         assert_eq!(
-            SigningKeyId::try_from("signed_curve25519:Abc-1").unwrap_err(),
+            ServerSigningKeyId::try_from("signed_curve25519:Abc-1").unwrap_err(),
             Error::UnknownKeyAlgorithm,
         );
     }
@@ -92,7 +96,7 @@ mod tests {
     #[test]
     fn missing_delimiter() {
         assert_eq!(
-            SigningKeyId::try_from("ed25519|Abc_1").unwrap_err(),
+            ServerSigningKeyId::try_from("ed25519|Abc_1").unwrap_err(),
             Error::MissingSigningKeyDelimiter,
         );
     }